CRenderCapture: split into separate classes
