+++ xbmc/cores/VideoPlayer/VideoRenderers/CMakeLists.txt
@@ -24,9 +24,11 @@ set(HEADERS BaseRenderer.h
 
 if(CORE_SYSTEM_NAME STREQUAL windows OR CORE_SYSTEM_NAME STREQUAL windowsstore)
   list(APPEND SOURCES WinRenderer.cpp
-                      OverlayRendererDX.cpp)
+                      OverlayRendererDX.cpp
+                      RenderCaptureDX.cpp)
   list(APPEND HEADERS WinRenderer.h
-                      OverlayRendererDX.h)
+                      OverlayRendererDX.h
+                      RenderCapture.h)
 endif()
 
 if (OPENGL_FOUND OR OPENGLES_FOUND)
@@ -36,9 +38,11 @@ endif()
 
 if(OPENGL_FOUND)
   list(APPEND SOURCES LinuxRendererGL.cpp
-                      FrameBufferObject.cpp)
+                      FrameBufferObject.cpp
+                      RenderCaptureGL.cpp)
   list(APPEND HEADERS LinuxRendererGL.h
-                      FrameBufferObject.h)
+                      FrameBufferObject.h
+                      RenderCaptureGL.h)
 endif()
 
 if(OPENGLES_FOUND AND ("android" IN_LIST CORE_PLATFORM_NAME_LC OR
@@ -48,9 +52,11 @@ if(OPENGLES_FOUND AND ("android" IN_LIST CORE_PLATFORM_NAME_LC OR
                        "x11" IN_LIST CORE_PLATFORM_NAME_LC OR
                        "wayland" IN_LIST CORE_PLATFORM_NAME_LC))
   list(APPEND SOURCES LinuxRendererGLES.cpp
-                      FrameBufferObject.cpp)
+                      FrameBufferObject.cpp
+                      RenderCaptureGLES.cpp)
   list(APPEND HEADERS LinuxRendererGLES.h
-                      FrameBufferObject.cpp)
+                      FrameBufferObject.cpp
+                      RenderCaptureGLES.h)
 endif()
 
 core_add_library(videorenderers)
